requirements.txt
flask==2.2.5
werkzeug==2.2.3
flask-cors==4.0.0
flask-jwt-extended==4.5.3
flask-mongoengine==1.0.0
flask-bcrypt==1.0.1
cryptography==41.0.7
python-dotenv==1.0.0
flask-limiter==3.5.0
requests==2.31.0
textblob==0.17.1
mongoengine==0.27.0
pymongo<4.0.0
mongomock==4.1.2
pytest==7.4.3
